Add files via upload
Dive into a comprehensive exploration of prescription drug trends with our PowerBI visualization, sourced from the Controlled Substance Monitoring Database (CSMD) for the pivotal years 2018-2022. This project, courtesy of data from the TDH-office of Informatics & Analytics of the TN Department of Health, offers a meticulous breakdown of opioid prescription patterns. Beyond mere visualization, the project stands as a testament to the power of modern data processing, with an intricate ETL process employing Power Query and Pivot Tables to cleanse and model the data, ensuring accuracy and insightful analysis.